Move printing to separate file
